Wait for the TC flag when sending over UART
The UART has a buffer, and we may not have sent all of the bytes yet
when the writer function returns. This worked fine in post 2 when we
hung immediately afterwards, but when used in a panic! handler that
resets the MCU, it can cause bytes to be left untransmitted. Presumably
this is would also be a data race with disabling the UART after a
transmission.

The write function now waits for the TC, or "Transmit Complete" flag to
be set, indicating that the buffer has been emptied.