Make Pin and Gpio construction safer
This adds a locking mechanism to the Port struct to indicate which
pins are in use. This locking is a form of interior mutability.

The main reason for this is that we want to ensure that a Pin cannot
outlive its parent Port. This means each pin has to borrow the
Port. If we did that with mutable references, we could only be able to
access one Pin. This means that Pins must take immutable borrows to
their parent.

Since we take an immutable borrow for Pin construction, we must use
`AtomicBool` to ensure that multiple references to the same Port can't
clobber each other during Pin construction.

We also clean up the use of mutable pointers in Pins, and instead
properly wrap the hardware registers of the Port with an `UnsafeCell`.