Clean up watchdog and BSS initialization
BSS init was being done AFTER the sim was created, which meant that we
couldn't always be sure that the SIM_INIT lock was `false` at
startup. This moves bss init to a single unsafe block with the
watchdog at the top of `main`, so that all of the "housekeeping" is
done at once.